+++ Classes/Service/Shopware/AbstractShopwareApiClient.php
@@ -582,9 +582,15 @@ abstract class AbstractShopwareApiClient implements SingletonInterface, Abstract
      */
     public function findAll($doCacheRequest = true, $params = [])
     {
-        $shopwareModels = new ObjectStorage();
+        $shopwareModelArray = [];
+
+        $params['limit'] = isset($params['limit']) ? $params['limit'] : 1000;
+        $params['start'] = isset($params['start']) ? $params['start'] : 0;
+
+            // first request has default limit of 1000
         $result = $this->get($this->getValidEndpoint(), $params, $doCacheRequest);
         if ($result) {
+
             $token = (isset($result->pxShopwareTypo3Token)) ? (bool)$result->pxShopwareTypo3Token : false;
             if (isset($result->data) && is_array($result->data)) {
                 foreach ($result->data as $data) {
@@ -592,11 +598,47 @@ abstract class AbstractShopwareApiClient implements SingletonInterface, Abstract
                         /** @var \Portrino\PxShopware\Domain\Model\AbstractShopwareModel $shopwareModel */
                         $shopwareModel = $this->objectManager->get($this->getEntityClassName(), $data, $token);
                         if ($shopwareModel != null) {
-                            $shopwareModels->attach($shopwareModel);
+                            $shopwareModelArray[$shopwareModel->getId()] = $shopwareModel;
                         }
                     }
                 }
+                $total = $result->total;
+
+                    // shop has more items than first request returned? poll again!
+                if ($total > 0 && $total > count($shopwareModelArray) + $params['start']) {
+
+                    $i = 0;
+                        // safety break
+                    while ($i < 99) {
+                            // increase offset (called start in shopware)
+                        $params['start'] = $params['start'] + $params['limit'];
+                            // get API result
+                        $additionalResults = $this->findByParams($params, $doCacheRequest);
+                        foreach ($additionalResults as $additionalResult) {
+                                // add new items to original array, if not already there
+                            if (!array_key_exists($additionalResult->getId(), $shopwareModelArray)) {
+                                $shopwareModelArray[$additionalResult->getId()] = $additionalResult;
+                            }
+                        }
+
+                            // stop if API returns empty
+                        if ($additionalResults->count() == 0) {
+                            break;
+                        }
+                            // stop if total count is reached
+                        if (count($shopwareModelArray) >= $total) {
+                            break;
+                        }
+                        $i++;
+                    }
+                }
             }
+        }
+
+            // transform array of models to ObjectStorage
+        $shopwareModels = new ObjectStorage();
+        foreach ($shopwareModelArray as $shopwareModel) {
+            $shopwareModels->attach($shopwareModel);
         }
         return $shopwareModels;
     }