+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # PxShopware Change log
 
+3.0.1 - 2017-01-24
+------------------
+* [BUGFIX] updates AbstractShopwareApiClient->findAll to allow more than 1000 items (shopware api default limit)
+* [BUGFIX] change EAN field from _stringS to _textS
+
 3.0.0 - 2016-11-09
 ------------------
 **TYPO3 6.2 Compatibility Release**